UpgradedVehicles v3.2.0 - Consolidated files - Using calculated execution location for config file
