--- UpgradedVehicles/Patchers/DamageSystem_Patcher.cs
@@ -1,28 +1,0 @@
-﻿namespace UpgradedVehicles.Patchers
-{
-    using Harmony;
-    using UnityEngine;
-
-    [HarmonyPatch(typeof(DamageSystem))]
-    [HarmonyPatch(nameof(DamageSystem.CalculateDamage))]
-    internal class DamageSystem_CalculateDamage_Patch
-    {
-        [HarmonyPostfix]
-        internal static void Postfix(ref float __result, GameObject target)
-        {
-            var vehicle = target.GetComponent<Vehicle>();
-
-            if (vehicle != null) // Target is vehicle
-            {
-                var vehicleUpgrader = VehicleUpgrader.GetUpgrader(vehicle);
-
-                if (vehicleUpgrader == null)
-                {
-                    return;
-                }
-
-                __result = vehicleUpgrader.GeneralDamageReduction * __result;
-            }
-        }
-    }
-}
