--- UpgradedVehicles/Patchers/Exosuit_Patcher.cs
@@ -1,15 +1,0 @@
-﻿namespace UpgradedVehicles.Patchers
-{
-    using Harmony;
-
-    [HarmonyPatch(typeof(Exosuit))]
-    [HarmonyPatch(nameof(Exosuit.OnUpgradeModuleChange))]
-    internal class Exosuit_OnUpgradeModuleChange_Patcher
-    {
-        [HarmonyPostfix]
-        internal static void Postfix(ref Exosuit __instance, TechType techType)
-        {
-            VehicleUpgrader.GetUpgrader(__instance)?.UpgradeVehicle(techType);
-        }
-    }
-}
