--- UpgradedVehicles/Patchers/SeaMoth_Patcher.cs
@@ -1,15 +1,0 @@
-﻿namespace UpgradedVehicles.Patchers
-{
-    using Harmony;
-
-    [HarmonyPatch(typeof(SeaMoth))]
-    [HarmonyPatch(nameof(SeaMoth.OnUpgradeModuleChange))]
-    internal class SeaMoth_OnUpgradeModuleChange_Patcher
-    {
-        [HarmonyPostfix]
-        internal static void Postfix(ref SeaMoth __instance, TechType techType)
-        {
-            VehicleUpgrader.GetUpgrader(__instance)?.UpgradeVehicle(techType);
-        }
-    }
-}
