+++ UpgradedVehicles/UpgradedVehicles.csproj
@@ -63,10 +63,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Craftables\SpeedBooster.cs" />
-    <Compile Include="Patchers\DamageSystem_Patcher.cs" />
-    <Compile Include="Patchers\Exosuit_Patcher.cs" />
-    <Compile Include="Patchers\SeaMoth_Patcher.cs" />
+    <Compile Include="SpeedBooster.cs" />
+    <Compile Include="Patchers.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QPatch.cs" />
     <Compile Include="SaveData\BonusSpeedStyles.cs" />
