+++ UpgradedVehicles/mod.json
@@ -2,10 +2,10 @@
   "Id": "UpgradedVehicles",
   "DisplayName": "UpgradedVehicles",
   "Author": "PrimeSonic",
-  "Version": "3.1.1",
-  "Requires": [ ],
+  "Version": "3.2.0",
   "Enable": true,
   "AssemblyName": "UpgradedVehicles.dll",
-  "EntryMethod": "UpgradedVehicles.QPatch.Patch",
-  "Config": {}
+  "Game": "Subnautica",
+  "Dependencies": [ "SMLHelper" ],
+  "EntryMethod": "UpgradedVehicles.QPatch.Patch"
 }
